[ci] Basic cli and deamon tests
We can easily isolate and test cli.
However for most parts we can't isolate and test daemon only. In order to cope with that problem we first test cli, and then we test daemon+cli. Therefore it will follow that if errors arise in cli only part then there would be an error in cli part, if errors arise in daemon+cli part and not in cli part then there would be an error in daemon.