[bash/daemon] Capture bulitns in a new way
Capture output of bulitns in a new way.
Now we redirect output of builtins to new in-memory temporary files, and then we read the memory files using cat, so the output of builtin will appear to nhi as belonging to new process.
In that way builtins are executed in the same shell, not in new shells (like in the case of bash pipes)